vacinometro-scraping-api/requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
Flask==2.1.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
outcome==1.1.0
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
wsproto==1.1.0
zipp==3.8.0
